Backend/requirements.txt
Flask
Flask-Cors
Flask-Login
Flask-SQLAlchemy
python-dotenv
reportlab
requests
